feat: Add online user functionality to chat
The code changes in the backend and frontend implement the functionality to track online users in the chat application. When a user joins a room, their user ID is added to the `onlineUsers` object. When a user disconnects, their user ID is removed from the `onlineUsers` object. The `onlineUsers` object is then emitted to all connected clients using the `onlineUsers` event. This commit adds the necessary code to achieve this functionality.